requirements.txt
streamlit
pydantic
weaviate
loguru
pymongo
python-dotenv
openai
networkx
scikit-learn
tiktoken